backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
langgraph==0.2.40
pydantic==2.9.2
numpy>=1.26.0,<2.0.0
pandas==2.2.2
scikit-learn==1.5.1
httpx==0.27.2
python-multipart==0.0.9
aiofiles==24.1.0
reportlab==4.2.2
datasets==2.18.0
huggingface-hub>=0.23.0

